fixed ci
